+++ include/open62541pp/types/Builtin.h
@@ -15,6 +15,7 @@ namespace opcua {
  */
 class String : public TypeWrapper<UA_String, UA_TYPES_STRING> {
 public:
+    // NOLINTNEXTLINE, false positive?
     using TypeWrapperBase::TypeWrapperBase;  // inherit contructors
 
     explicit String(std::string_view str);
@@ -28,6 +29,7 @@ public:
  */
 class Guid : public TypeWrapper<UA_Guid, UA_TYPES_GUID> {
 public:
+    // NOLINTNEXTLINE, false positive?
     using TypeWrapperBase::TypeWrapperBase;  // inherit contructors
 
     Guid(UA_UInt32 data1, UA_UInt16 data2, UA_UInt16 data3, std::array<UA_Byte, 8> data4);