+++ include/open62541pp/types/DataValue.h
@@ -2,6 +2,7 @@
 
 #include <cstdint>
 #include <optional>
+#include <utility>  // forward
 
 #include "open62541pp/TypeWrapper.h"
 #include "open62541pp/open62541.h"
@@ -28,6 +29,16 @@ public:
         std::optional<uint16_t> serverPicoseconds,
         std::optional<UA_StatusCode> statusCode
     );
+
+    /// Create Variant from scalar value.
+    /// @see Variant::fromScalar
+    template <typename... Args>
+    static DataValue fromScalar(Args&&... args);
+
+    /// Create Variant from array.
+    /// @see Variant::fromArray
+    template <typename... Args>
+    static DataValue fromArray(Args&&... args);
 
     /// Get value as pointer (might be `nullptr` if not set).
     Variant* getValuePtr() noexcept;
@@ -62,5 +73,23 @@ public:
     /// Set status code.
     void setStatusCode(UA_StatusCode statusCode);
 };
+
+/* --------------------------------------- Implementation --------------------------------------- */
+
+template <typename... Args>
+DataValue DataValue::fromScalar(Args&&... args) {
+    DataValue dv{};
+    asWrapper<Variant>(dv->value) = Variant::fromScalar(std::forward<Args>(args)...);
+    dv->hasValue = true;
+    return dv;
+}
+
+template <typename... Args>
+DataValue DataValue::fromArray(Args&&... args) {
+    DataValue dv{};
+    asWrapper<Variant>(dv->value) = Variant::fromArray(std::forward<Args>(args)...);
+    dv->hasValue = true;
+    return dv;
+}
 
 }  // namespace opcua