+++ CHANGELOG.md
@@ -11,8 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `asWrapper` function to cast native refs to wrapper refs (#30, #31)
 - `DataValue::getValuePtr` method
+- Static methods `DataValue::fromScalar` and `DataValue::fromArray`
 - `Server::getNamespaceArray` method
 - `Server::runIterate` method, e.g. to run server in existing event loop
+- Generic `services::readAttribute` and `services::writeAttribute` functions
 
 ### Changed
 
@@ -20,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Return reference from `ExpandedNodeId::getNodeId`
 - Pass `LocalizedText` instead of members (`locale`, `text`) to `Node::writeDisplayName`, `Node::writeDescription`, `services::writeDisplayName`, `services::writeDescription` (#29)
 - Remove `Server::getConfig` method
+- Rename `Node::writeModellingRule` -> `Node::addModellingRule`
 
 ### Fixed